Сделал один общий модуль управления модальными окнами
