+++ js/utils/random-values.js
@@ -26,7 +26,4 @@ const createIdGenerator = (min, max, uniqueness = false) => {
 };
 
 
-const isEscapeKey = (evt) => evt.key === 'Escape';
-
-
-export { getRandomInteger, getRandomElement, createIdGenerator, isEscapeKey };
+export { getRandomInteger, getRandomElement, createIdGenerator };
